malloc should take a size_t, you can't allocate more than that anyway

+++ native/modload/loaded_mods.h
@@ -10,7 +10,7 @@
 #include "p3a/p3a.h"
 
 namespace SenLib::ModLoad {
-using PMalloc = void* (*)(uint64_t size);
+using PMalloc = void* (*)(size_t size);
 using PFree = void (*)(void* memory);
 
 struct P3AData {
